+++ src/FileUpload/FileUpload.ts
@@ -1,6 +1,6 @@
-import { CompleteMultipartUpload } from '../MultipartUpload/CompleteMultipartUpload'
-import { InitiateMultipartUpload } from '../MultipartUpload/InitiateMultipartUpload'
-import { DeleteMultipartUpload } from '../MultipartUpload/DeleteMultipartUpload'
+import { CompleteMultipartUpload } from '../AwsRequest/MultipartUpload/CompleteMultipartUpload'
+import { InitiateMultipartUpload } from '../AwsRequest/MultipartUpload/InitiateMultipartUpload'
+import { DeleteMultipartUpload } from '../AwsRequest/MultipartUpload/DeleteMultipartUpload'
 import { ResumeInterruptedUpload } from '../AwsRequest/ResumeInterruptedUpload'
 import { PutPart } from '../AwsRequest/PutPart'
 import { ReuseS3Object } from './ReuseS3Object'
